#86: moved TemplateResponse to LegacyConversationWebservice; other minor changes
